Move estimated gas time into its own hook
